--- a/devito/dse/extended_sympy.py
+++ b/devito/dse/extended_sympy.py
@@ -39,6 +39,28 @@ class Add(sympy.Add, UnevaluatedExpr):
     pass
 
 
+class taylor_sin(TrigonometricFunction):
+
+    """
+    Approximation of the sine function using a Taylor polynomial.
+    """
+
+    @classmethod
+    def eval(cls, arg):
+        return 0.0 if arg == 0.0 else eval_taylor_sin(arg)
+
+
+class taylor_cos(TrigonometricFunction):
+
+    """
+    Approximation of the cosine function using a Taylor polynomial.
+    """
+
+    @classmethod
+    def eval(cls, arg):
+        return 1.0 if arg == 0.0 else eval_taylor_cos(arg + 1.5708)
+
+
 class bhaskara_sin(TrigonometricFunction):
 
     """
@@ -65,3 +87,11 @@ class bhaskara_cos(TrigonometricFunction):
 
 def eval_bhaskara_sin(angle):
     return 16.0*angle*(3.1416-abs(angle))/(49.3483-4.0*abs(angle)*(3.1416-abs(angle)))
+
+
+def eval_taylor_sin(angle):
+    return angle-(angle*angle*angle/6.0*(1.0-angle*angle/20.0))
+
+
+def eval_taylor_cos(angle):
+    return 1.0-.5*angle*angle*(1.0-angle*angle/12.0)