--- a/devito/dse/graph.py
+++ b/devito/dse/graph.py
@@ -23,7 +23,7 @@ from collections import OrderedDict, namedtuple
 
 from sympy import (Eq, Indexed)
 
-from devito.dse.inspection import terminals
+from devito.dse.inspection import is_time_invariant, terminals
 
 __all__ = ['temporaries_graph']
 
@@ -42,10 +42,12 @@ class Temporary(Eq):
     def __new__(cls, lhs, rhs, **kwargs):
         reads = kwargs.pop('reads', [])
         readby = kwargs.pop('readby', [])
+        time_invariant = kwargs.pop('time_invariant', False)
         scope = kwargs.pop('scope', 0)
         obj = super(Temporary, cls).__new__(cls, lhs, rhs, **kwargs)
         obj._reads = set(reads)
         obj._readby = set(readby)
+        obj._is_time_invariant = time_invariant
         obj._scope = scope
         return obj
 
@@ -56,6 +58,10 @@ class Temporary(Eq):
     @property
     def readby(self):
         return self._readby
+
+    @property
+    def is_time_invariant(self):
+        return self._is_time_invariant
 
     @property
     def is_terminal(self):
@@ -119,16 +125,17 @@ def temporaries_graph(temporaries, scope=0):
     """
 
     mapper = OrderedDict()
-    Node = namedtuple('Node', ['rhs', 'reads', 'readby'])
+    Node = namedtuple('Node', ['rhs', 'reads', 'readby', 'time_invariant'])
 
     for lhs, rhs in [i.args for i in temporaries]:
         reads = {i for i in terminals(rhs) if i in mapper}
-        mapper[lhs] = Node(rhs, reads, set())
+        mapper[lhs] = Node(rhs, reads, set(), is_time_invariant(rhs, mapper))
         for i in mapper[lhs].reads:
             assert i in mapper, "Illegal Flow"
             mapper[i].readby.add(lhs)
 
-    nodes = [Temporary(k, v.rhs, reads=v.reads, readby=v.readby, scope=scope)
+    nodes = [Temporary(k, v.rhs, reads=v.reads, readby=v.readby,
+                       time_invariant=v.time_invariant, scope=scope)
              for k, v in mapper.items()]
 
     return OrderedDict([(i.lhs, i) for i in nodes])