… devito/interfaces.py
… devito/logger.py
… devito/propagator.py