This PR introduces:

* the concept of ``DSE transformation step`` (see the ``dse_transformation`` decorator in symbolics.py)
* a complete new DSE transformation step: code motion. This is implemented in ``_replace_time_invariants``. Time-invariant sub-expressions are sought and given a score. If the score is greater than an empirically determined threshold, then the sub-expression at hand is "lifted" outside the time loop and evaluated in a temporary
* taylor-based approximation of sin/cos. The default choice, however, is still Bhaskara, as tests start failing with Taylor. I wonder if this is due to the remarkably bad approximation that Taylor gives as one gets far from default center, 0
* refactoring and improvements of various nature for the DSE
* extended propagator to emit time-invariant code outside of the time loop
* new tests

I run some tests to verify the impact of the new transformation (single socket, 8 threads, icc O2 xHost, pinning, etc):

* TTI space order 4: runtime drops from 127s to 91s (~1.4X speed-up)
* TTI space order 10: runtime drops from 373s to 322s (1.16X speed-up)

If *no* code motion is performed and sin/cos are replaced with the respective Bhaskara approximations, the execution time in TTI space order 4 only drops from 127s to 105s. 
If code motion is performed and *no* Bhaskara approximation is introduced, than the run-time is pretty much the same (92s in TTI space order 4) *but* the compile time is much higher (icc takes more to vectorize loops with sin/cos). For example, in TTI space order 10, the compilation time drops from 74s to 16s thanks to Bhaskara