@mloubout 2017-01-03
Even if it doesn't really matter, the point is we keep running test with this orders that should not be used. So evenso it does not impact your result, it means noone will remember it should not be used if people don't make sure to use it properly wether or not it is for a proper geophysical application. I don't want to ebd up with shotty results in a few month because everyone forgot it again.