@FabioLuporini 2017-01-03
Yes, so it's best if we add checks at the beginning of TTI which prevent the execution of what u believe should never be run