@ggorman 2017-01-03
I'm supporting @mloubout's point here. It feels like we are leaving behind a landmine for someone to step on...

How about making use of the python warnings module?