@FabioLuporini 2017-01-04
I also agree with his point, that's why I proposed to add checks for when one tries to run TTI