@mloubout 2017-01-05
Another thing. There is with that PR, no possibilty to disable dse. This is sometimes required for non implemented case (I sended you a private message withone case ). Can we add a `no-dse` case to have everything covered?