@FabioLuporini 2017-01-14
@mloubout If you instantiate operators passing ``dse=None``, no DSE transformations is applied