@mlange05 2017-01-16
OK, so basically this looks really great - although a little complicated for novices. ;) Tying the allocation of temporaries directly into the Propagator worries me slightly, since it makes the 2.0 re-write harder, but in the interest of time I think we can merge this now and unpick later. Also, the TTI space order discussion is orthogonal to this PR, so I think this is good to go.