@FabioLuporini 2017-01-16
Thanks for merging this. The reason temporaries are allocated in the propagator is that the DSE is orthogonal to code generation, as it's just pure expression manipulation. 

I think we have to be prepared in 2.0 to create a general mechanism to allocate temporaries at a certain loop nest depth. With such a mechanism, allocation of temporaries for stencils, additional expressions (eg sparse point interpolation), time invariants etc would be handled uniformly with minimal pain