@mlange05 2017-01-16
I agree, I was just pointing out that the "allocation" code used for generic FT on symbols and for temporaries should all come from the same place.

Also, I just realised this, but are we actually freeing temporaries in the kernel too, or are we leaking memory here?