@FabioLuporini 2017-01-16
Good spot. I'm 100% I handled that, and in fact now that I look at the code, it's missing other stuff. I must have dropped a commit during an interactive rebase.

I'm rebasing #170 to include what's been shamefully dropped here, so that (amongst the other things missing) we don't leak memory